refactor: Product 파일 분리
